Ajout d'un fichier concern
